@@ -27,9 +27,8 @@
     "@bower_components/leaflet-utfgrid": "danzel/Leaflet.utfgrid",
     "@bower_components/leaflet.fullscreen": "brunob/leaflet.fullscreen",
     "@elastic/filesaver": "^1.1.2",
+    "antd": "^5.16.1",
     "async": "^2.0.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-remove-strict-mode": "^0.0.2",
     "body-parser": "^1.16.1",
     "bootstrap": "^3.3.6",
     "c3": "^0.4.18",
@@ -46,9 +45,8 @@
     "express-expose": "^0.3.4",
     "express-session": "^1.15.1",
     "geopoint": "^1.0.1",
-    "griddle-react": "^0.6.1",
+    "griddle-react": "^0.8.1",
     "hamljs": "^0.6.2",
-    "hiredis": "^0.5.0",
     "http-proxy": "~1.8.1",
     "jquery": "~3.3.1",
     "jquery-ui": "^1.12.1",
@@ -66,12 +64,12 @@
     "pg": "^4.5.6",
     "pouchdb": "^7.0.0",
     "prop-types": "^15.5",
-    "react": "^15.5",
+    "react": "^17.0",
     "react-addons-css-transition-group": "^15.5",
     "react-addons-pure-render-mixin": "^15.5",
     "react-c3js": "^0.1.19",
     "react-datetime": "^2.11.0",
-    "react-dom": "^15.5",
+    "react-dom": "^17.0",
     "reactify": "^1.1.1",
     "redis": "^2.6.2",
     "request": "^2.73.0",
@@ -81,19 +79,20 @@
     "tablesorter": "^2.29.0",
     "underscore": "^1.8.3",
     "winston": "^2.3.1",
-    "wurl": "^2.5.2"
+    "wurl": "^2.5.2",
+    "yarn": "^1.22.19"
   },
   "devDependencies": {
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "7.0.0",
     "assert": "~1.3.0",
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.23.1",
-    "babel-jest": "^18.0.0",
-    "babel-plugin-module-resolver": "^2.5.0",
+    "babel-plugin-module-resolver": "3.0.0",
     "babel-plugin-transform-promise-to-bluebird": "^1.1.1",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.18.0",
-    "babelify": "^7.3.0",
+    "babelify": "10.0.0",
     "browserify": "^14.5.0",
     "browserify-css": "^0.14.0",
     "browserify-shim": "^3.8.14",
@@ -104,13 +103,13 @@
     "eslint-plugin-import": "^2.2.0",
     "fancy-log": "^1.3.2",
     "fs-extra": "^7.0.1",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^7.0.0",
+    "gulp": "3.9.1",
+    "gulp-babel": "^8.0.0",
     "gulp-less": "^3.3.2",
     "gulp-sourcemaps": "^2.6.1",
     "gulp-tap": "^1.0.1",
     "gulp-uglify": "^3.0.0",
-    "jest": "^18.0.0",
+    "jest": "^24.0.0",
     "npm-shrinkwrap": "200.5.1",
     "react-test-renderer": "^15.5",
     "reactify": "^1.1.1",
@@ -120,6 +119,9 @@
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
     "watchify": "^3.7.0"
+  },
+  "resolutions": {
+    "graceful-fs": "^4.2.11"
   },
   "keywords": [],
   "repository": "git://github.com/idigbio/idigbio-portal",
@@ -138,7 +140,9 @@
     "moduleNameMapper": {
       "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
     },
-    "setupTestFrameworkScriptFile": "<rootDir>/__tests__/init"
+    "setupFilesAfterEnv": [
+      "<rootDir>/__tests__/init"
+    ]
   },
   "browserify": {
     "transform": [
