@@ -314,7 +314,7 @@ module.exports = function(elid, options){
                     [dwc['dwc:scientificName'],n,'No Name']
                 )
                 
-                var inf = ['<b><span class="record-count">'+helpers.formatNum(ind+offset+1)+'</span><a class="record-link" target="'+item.uuid+'" href="//www.idigbio.org/portal/records/'+item.uuid+'">View Record</a></b>'];
+                var inf = ['<b><span class="record-count">'+helpers.formatNum(ind+offset+1)+'</span><a class="record-link" target="'+item.uuid+'" href="/portal/records/'+item.uuid+'">View Record</a></b>'];
                 _.each(['genus','specificepithet','scientificname','country','stateprovince','lat','lon','institutioncode','collectioncode','catalognumber','datecollected'],function(term){
                     if(_.has(dwc,fields.byTerm[term].dataterm)){
                         inf.push(
